+++ src/compas_sandbox/equilibrium/__init__.py
@@ -1,4 +1,6 @@
 from .cra_pyomo import cra_solve
+from .cra_native import cra_solve_native
+from .cra_nlp import cra_problem
 from .cra_penalty_pyomo import cra_penalty_solve
 from .rbe_pyomo import rbe_solve
 from .cra_helper import (
@@ -25,6 +27,8 @@ from .pyomo_helper import (
 
 __all__ = [
     "cra_solve",
+    "cra_solve_native",
+    "cra_problem",
     "cra_penalty_solve",
     "rbe_solve",
     "equilibrium_setup",